packages/backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fastset-backend"
version = "0.1.0"
description = "FastSet BI Platform Backend"
requires-python = ">=3.8"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.116.1",
    "sqlalchemy>=2.0.43",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    # "pytest>=7.0.0",
    # "pytest-asyncio>=0.21.0",
    # "httpx>=0.25.0",
    # "pytest-cov>=4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/backend"]

